Workaround for adding PPA under debian.
Adding PPA repositories does not seem to be supported under Debian
currently. See the following links for other people with the same issue:

* https://github.com/ansible/ansible/issues/21766
* https://github.com/ansible/ansible/issues/20345
* https://github.com/ansible/ansible/issues/13350

This commit adds an alternative for debian that imports the key and adds
the repository manually. Currently, the codename is hardcoded to xenial
(which must be an ubuntu codename), but I do not know which ubuntu
codenames correspond to which debian versions...